Continue running repairs whilst we are making active progress
The current control we have for repairs timing out is a bit too coarse.
It imposes a global level timeout so even if we are making progress, if
we hit the timeout, the segment gets cut off. Given the timeout is to
detect hanging repairs, we can be smarter and have a renewable ticker
with indication that after some time of no progress, we cut it off

Our sign of progress is fed in from Cassandra itself so if it deems itself
making progress, that's good enough for us.