Avoid aborting not started repairs
Reaper aborts on nodes that have seemingly not started repairs, we're
going to lean on the other end and assume that a non-started repair
simply does not exist.

If a segment is running subsequently, it'll be picked up on a subsequent
check of the Repair stats, so we aren't risking much here.